Prepare before the major rewrite (#555)

+++ bot/libs/utils/context.py
@@ -1,60 +1,39 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Optional
+from typing import TYPE_CHECKING, Optional
 
 import discord
 from discord.ext import commands
 
-from .utils import produce_error_embed
+from .view import KumikoView
 
 if TYPE_CHECKING:
-    from bot.kumikocore import KumikoCore
+    from bot.kumiko import Kumiko
 
 NO_CONTROL_MSG = "This view cannot be controlled by you, sorry!"
+
 
 # Why not subclass KumikoView?
 # It results in a circular logic, so instead
 # we subclassed discord.ui.View and pretty much implement what KumikoView does anyways
-class ConfirmationView(discord.ui.View):
-    def __init__(self, ctx: KContext, timeout: float, delete_after: bool) -> None:
-        super().__init__(timeout=timeout)
-        self.ctx = ctx
+class ConfirmationView(KumikoView):
+    def __init__(self, ctx, timeout: float, delete_after: bool):
+        super().__init__(ctx, timeout=timeout)
         self.value: Optional[bool] = None
-        self.delete_after: bool = delete_after
+        self.delete_after = delete_after
         self.message: Optional[discord.Message] = None
 
-    async def interaction_check(self, interaction: discord.Interaction, /) -> bool:
-        if interaction.user and interaction.user.id in (
-            self.ctx.bot.application.owner.id,  # type: ignore
-            self.ctx.author.id,
-        ):
-            return True
-
-        await interaction.response.send_message(NO_CONTROL_MSG, ephemeral=True)
-        return False
-
-    async def on_error(
-        self,
-        interaction: discord.Interaction,
-        error: Exception,
-        item: discord.ui.Item[Any],
-        /,
-    ) -> None:
-        await interaction.response.send_message(
-            embed=produce_error_embed(error), ephemeral=True
-        )
-        self.stop()
-
     async def on_timeout(self) -> None:
-        if self.message:
-            if self.delete_after:
-                await self.message.delete()
-                return
+        if self.delete_after and self.message:
+            await self.message.delete()
+        elif self.message:
             await self.message.edit(view=None)
 
-    async def delete_response(self, interaction: discord.Interaction) -> None:
+    async def delete_response(self, interaction: discord.Interaction):
         await interaction.response.defer()
-        await interaction.delete_original_response()
+        if self.delete_after:
+            await interaction.delete_original_response()
+
         self.stop()
 
     @discord.ui.button(
@@ -64,30 +43,32 @@ class ConfirmationView(discord.ui.View):
     )
     async def confirm(
         self, interaction: discord.Interaction, button: discord.ui.Button
-    ):
+    ) -> None:
         self.value = True
-        if self.delete_after:
-            await self.delete_response(interaction)
+        await self.delete_response(interaction)
 
     @discord.ui.button(
         label="Cancel",
         style=discord.ButtonStyle.red,
         emoji="<:redTick:596576672149667840>",
     )
-    async def cancel(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def cancel(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ) -> None:
         self.value = False
-        await self.delete_response(interaction)
+        await interaction.response.defer()
+        await interaction.delete_original_response()
+        self.stop()
 
 
-class KContext(commands.Context):
+class KumikoContext(commands.Context):
     """Kumiko's custom `commands.Context` with extra features"""
 
-    bot: KumikoCore
+    bot: Kumiko
 
     def __init__(self, **kwargs) -> None:
         super().__init__(**kwargs)
         self.pool = self.bot.pool
-        self.redis_pool = self.bot.redis_pool
         self.session = self.bot.session
 
     async def prompt(
@@ -136,8 +117,8 @@ class KContext(commands.Context):
         return members[0]
 
 
-class GuildContext(KContext):
-    """An `KContext` that represents a context found in a guild command"""
+class GuildContext(KumikoContext):
+    """An `KumikoContext` that represents a context found in a guild command"""
 
     author: discord.Member
     guild: discord.Guild
